+++ src/main/java/database/transaction/TransactionExecutor.java
@@ -1,7 +1,6 @@
-package database.connection;
+package database.transaction;
 
-import database.dao.TransactionCallable;
-import database.dao.TransactionRunnable;
+import database.context.ConnectionContext;
 
 import java.sql.Connection;
 import java.sql.SQLException;
